Hack for pinky40
There is a merger creating > 10^4 potential locations for the free ends,
so we need to switch to a less intensive method of finding the free
ends. Instead of starting from supervoxels next to the shaft, we start
from the shaft. The results are slightly different, lower the NRI by 1%
on the golden cube.