Mean affinity agglomeration in chunks
The code reads in a list of supervoxels and freeze every edges
containing them. Processing on the sorted list of edges, the
neighbouring supervoxels be added to the list and edges related to those
supervoxels will be frozen as well.

At this moment the inputs are generated by the region graph code, the
list of supervoxels are the supervoxels of the incomplete edges. After
finishing the chunkwise mean affinity agglomeration (using the chunk
region graphs). We complete all the edges. The final step is to put the
newly completed edges together with the frozen edges and the edges left
over through a big (global) agglomeration again.

Once we have the ability to keep track of the partially completed and
completed supervoxels after we put a subset of chunks to form a big
chunk, we can proceed with the octree structure.