Attach axons to their soma
For each axon free ends, if we cannot find a matching axon, try other
supervoxles touching the free ends. If we find another free ends (not
necessarily belonging to a axon), we will consider agglomerating it with
the target axon. If we find multiple matches, we pick the one with the
largest mean affinity. This change is mainly for attaching axons to
their soma, but also matches some spines to their dendrites.