Save the edges in the heap nodes
In order to be efficient, we sort the region graph and replace the nodes
of the edges with local ids, then write the updated region graph to the
hard drive in the same region graph file. Then load the region graph
again to populate the heap, this way we can free the memory of the
vectors used to sort the region graph, without keeping multiple copies
of the region graph in memory.