Use the boundary bit
The old code gives correct supervoxel count in the non-overlapped mode
because the boundary supervoxels will not be agglomerated until they are
no longer at the boundary. This is not true anymore when we process
chunks after overlapped agglomeration. Add a #ifdef to change the behavior
of the code to count it correctly.

This is really an ugly hack and a proper fix should be extra code in the
merge_chunks part to make the behavior in the non-overlapped mode the
same as the overlapped mode.