Add more flags to optimize for different stage
Extra flag was applied at the overlapping phase and beyond, which allows
the agglomeration of non local-maximal edges

Final was applied after we finish the overlapping phase. It will not
preserve the segid. It should be OK to apply it to the whole process,
but probably will not make a big difference and make debugging harder